FAPI: Enable usage of existing NV indexes for Fapi_NvCreate
Fapi_NvCreate can now create FAPI objects in keystore for existing
NV indexes unless the NV index is defined with a policy which does
not match the policy defined by Fapi_CreateNv.
If a policy is defined but the attributes NV_AUTHWRITE and
NV_AUTHREAD are set the object will be created without
a policy in keystore.
In other cases with a policy Fapi_NvCreate will return
TSS2_FAPI_RC_BAD_VALUE.

Signed-off-by: Juergen Repp <juergen_repp@web.de>